Added gameplay exception for badChoice
if user does not input Rock, Paper, or Scissors, the playGame function returns "No Score - Try Again".

This will possible change as I add the Scoring component, since it seems like a good choice will be to covert the returns of the playGame function to numerical values for scoring. I'm still unsure if it's the right move to create variables from the returned strings that hold the numerical value.